% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genoprob_gbs.R
\name{calc_genoprob_gbs}
\alias{calc_genoprob_gbs}
\title{Calculate genotype probabilities from GBS allele counts}
\usage{
calc_genoprob_gbs(
  counts,
  map,
  error_prob1 = 0.002,
  error_prob2 = 0.002,
  map_function = c("haldane", "kosambi", "c-f", "morgan"),
  cores = 1
)
}
\arguments{
\item{counts}{Three-dimensional array of counts, positions x individuals x two alleles, with alleles A and B
where we are expecting genotypes AA and AB}

\item{map}{Vector of marker positions, same length as \code{nrow(counts)}, in cM}

\item{error_prob1}{Error probability for sequencing errors}

\item{error_prob2}{Error probability for locus errors}

\item{map_function}{Map function for converting from cM distances to recombination fractions}

\item{cores}{Number of CPU cores to use, for multi-core calculations}
}
\value{
Matrix of genotype probabilities, positions x individuals, for Pr(het)
}
\description{
Calculate genotype probabilities using an HMM, from high-throughput sequencing data with counts of alleles at SNPs
}
