% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabXO.R
\name{grab2XO}
\alias{grab2XO}
\title{Grab all double-crossover locations}
\usage{
grab2XO(xoloc)
}
\arguments{
\item{xoloc}{List of matrices, of crossover locations, as output from \code{\link[=inferXOloc]{inferXOloc()}}.}
}
\value{
Matrix with locations of pair of crossovers, when there are exactly two
}
\description{
Grab all double-crossover locations, as a matrix with two columns, for cases of exactly 2 crossovers
}
