% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabXO.R
\name{grabXO}
\alias{grabXO}
\title{Grab all crossover locations}
\usage{
grabXO(xoloc)
}
\arguments{
\item{xoloc}{List of matrices, of crossover locations, as output from \code{\link[=inferXOloc]{inferXOloc()}}.}
}
\value{
Vector of crossover locations
}
\description{
Grab all crossover locations, as a vector
}
