% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferXOloc.R
\name{inferXOloc}
\alias{inferXOloc}
\title{Infer crossover locations}
\usage{
inferXOloc(genoprob, map, low = 0.1, high = 0.9)
}
\arguments{
\item{genoprob}{Matrix of genotype probabilities, positions x individuals, as produced by \code{\link[=calc_genoprob_gbs]{calc_genoprob_gbs()}}}

\item{map}{Vector of marker positions, same length as \code{nrow(genoprob)}}

\item{low}{Lower threshold; if probability below this threshold, infer homozygous}

\item{high}{Higher threshold; if probaability above this threshold, infer heterozygous}
}
\value{
List of matrices (of length \code{ncol(genoprob)}, each having
columns with estimated location and left and right interval
endpoints. If no crossovers, it will be a matrix with no rows.
}
\description{
Infer crossover locations from a matrix of genotype probabilities as produced by \code{\link[=calc_genoprob_gbs]{calc_genoprob_gbs()}}
}
