% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_ids.R
\name{order_ids}
\alias{order_ids}
\title{Get numeric order of individual IDs}
\usage{
order_ids(ids, decreasing = FALSE)
}
\arguments{
\item{ids}{Vector of individual IDs (as character string)}

\item{decreasing}{If TRUE, get decreasing order (largest to smallest)}
}
\value{
Input IDs, sorted numerically
}
\description{
Get numeric order of individual IDs, when they're of the form blah25 or blah25-2
}
\seealso{
\code{sort_ids()}
}
