% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randXOloc.R
\name{randXOloc}
\alias{randXOloc}
\title{Randomizing estimated crossover locations.}
\usage{
randXOloc(xoloc)
}
\arguments{
\item{xoloc}{Either a matrix, or a list of matrices, of crossover locations, as output from \code{\link[=inferXOloc]{inferXOloc()}}.}
}
\value{
Object like that input, but with estimated crossover locations randomized.
}
\description{
Randomizing estimated crossover
locations, uniform within their intervals.
}
