% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_counts.R
\name{read_counts}
\alias{read_counts}
\title{read counts data from a set of files}
\usage{
read_counts(files)
}
\arguments{
\item{files}{A vector of character strings with the counts data. These text files can be gzipped. Each file should be one backcross individual, and the files should have six whitespace-delimited columns: chromosome, basepairs positions, allele1, allele2, readcount1, readcount2, with allele1 being the allele for the homozygote parent. Alternatively, it can be a single directory, in which case we read all the .txt or .gz files in that directory.}
}
\value{
A list of data frames with the contents of the files
}
\description{
read counts data from a set of files
}
\details{
Names are taken from the file names...everything before ".txt" or ".gz", but removing "_read_counts" if it's part of the name.
So BCA81-2_read_counts.txt.gz becomes BCA81-2
}
