% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorg_counts.R
\name{reorg_counts}
\alias{reorg_counts}
\title{reorganize count data}
\usage{
reorg_counts(counts, map = NULL, clean_chr = TRUE, quiet = TRUE)
}
\arguments{
\item{counts}{A list of data frames, one per individual, with six columns: chromosome, basepairs
position, allele1, allele2, readcount1, and readcount 2, as read with \code{read_counts()}.}

\item{map}{Optional map of SNPs to consider in the output, as a
list of vectors of marker positions. If provided, any other
positions will be ignored.}

\item{clean_chr}{If TRUE, remove "chr" from the chromosome names, so chr13 becomes just 13}

\item{quiet}{If FALSE, print some tracing info}
}
\value{
A list containing \code{map} and \code{counts}, with the map being a list of vectors of Mbp positions, and the counts being a list of 3-dimensional arrays, position x individual x allele.
}
\description{
Reorganize count data, from by-individual data frames to a list of 3-dimensional arrays.
}
