% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_ids.R
\name{sort_ids}
\alias{sort_ids}
\title{Sort individual IDs numerically}
\usage{
sort_ids(ids, decreasing = FALSE)
}
\arguments{
\item{ids}{Vector of individual IDs (as character string)}

\item{decreasing}{If TRUE, sort in decreasing order (largest to smallest)}
}
\value{
Input IDs, sorted numerically
}
\description{
Sort individual IDs numerically, when they're of the form blah25 or blah25-2
}
\seealso{
\code{order_ids()}
}
