% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimXO.R
\name{trimXO}
\alias{trimXO}
\title{Trim tight double crossovers from crossover information}
\usage{
trimXO(xoloc, mind = 1)
}
\arguments{
\item{xoloc}{Either a matrix, or a list of matrices, of crossover locations, as output from \code{\link[=inferXOloc]{inferXOloc()}}.}

\item{mind}{Minimum allowed distance between crossovers}
}
\value{
Object like that input, but with double-crossovers within \code{mind} of each other removed
}
\description{
Trim tight double crossovers from crossover information
}
