% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoiGBS-package.R
\docType{package}
\name{xoiGBS-package}
\alias{xoiGBS}
\alias{xoiGBS-package}
\title{xoiGBS: Analyze Recombination and Crossover Interference with Sequence-Based Genotypes}
\description{
Analyze recombination and crossover interference from genotyping-by-sequencing (GBS) data on a backcross.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/xoiGBS}
  \item Report bugs at \url{https://github.com/kbroman/xoiGBS/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
